requirements.txt
argh==0.26.2
asgiref==3.5.2
brand==0.0.4
config2py==0.0.3
Django==4.0.4
djangorestframework==3.13.1
dol==0.1.84
importlib-resources==5.7.1
psycopg2-binary==2.9.3
py2store==0.1.11
pytz==2022.1
sqlparse==0.4.2
